test(worldgen): migrate VanillaGenerator-using tests to pumpkin-worldgen
* `pumpkin-worldgen/tests/biome.rs` — `biome_desert`,
  `wide_area_surface`, `sample_value`, `sample_multinoise_biome`.
  Originally in `pumpkin-world/src/biome/{mod,multi_noise}.rs`,
  disabled with `#[cfg(any())]` in the worldgen split because they
  instantiate `VanillaGenerator`. JSON fixtures are read via a small
  `asset()` helper that points at `pumpkin-world/assets/`.

* `pumpkin-worldgen/tests/chunk_generation.rs` —
  `generate_chunk_should_return`. End-to-end driver originally in
  `pumpkin-world/src/chunk_system/generation.rs`; needs
  `get_world_gen` (now in `pumpkin-worldgen`).

* Pure helper tests (`hash_seed_test`, `proper_network_bits_per_entry`)
  re-enabled in `pumpkin-world/src/biome/mod.rs` — they don't depend on
  `VanillaGenerator` and there was no reason to move them.

Verified: `cargo test -p pumpkin-worldgen --test biome --test chunk_generation`
→ 5 passed, 0 failed.
