+++ pumpkin-world/src/biome/mod.rs
@@ -48,100 +48,16 @@ pub fn hash_seed(seed: u64) -> i64 {
     i64::from_le_bytes(result[..8].try_into().unwrap())
 }
 
-// TODO: move biome tests to `pumpkin-worldgen` integration tests
-// (they depend on `VanillaGenerator` which now lives in that crate).
-#[cfg(any())]
+#[cfg(test)]
 mod test {
-    use pumpkin_data::{chunk::Biome, dimension::Dimension};
-    use pumpkin_util::read_data_from_file;
-    use serde::Deserialize;
+    use pumpkin_data::chunk::Biome;
 
-    use crate::{
-        ProtoChunk,
-        chunk::palette::BIOME_NETWORK_MAX_BITS,
-        generation::noise::router::multi_noise_sampler::{
-            MultiNoiseSampler, MultiNoiseSamplerBuilderOptions,
-        },
-    };
+    use crate::chunk::palette::BIOME_NETWORK_MAX_BITS;
 
-    use super::{BiomeSupplier, MultiNoiseBiomeSupplier, hash_seed};
+    use super::hash_seed;
 
-    #[test]
-    fn biome_desert() {
-        use crate::generation::generator::{GeneratorInit, VanillaGenerator};
-        use pumpkin_util::world_seed::Seed;
-        let seed = 13579;
-        let generator = VanillaGenerator::new(Seed(seed as u64), Dimension::OVERWORLD);
-        let multi_noise_config = MultiNoiseSamplerBuilderOptions::new(1, 1, 1);
-        let mut sampler =
-            MultiNoiseSampler::generate(&generator.base_router.multi_noise, &multi_noise_config);
-        let biome = MultiNoiseBiomeSupplier::OVERWORLD.biome(-24, 1, 8, &mut sampler);
-        assert_eq!(biome, &Biome::DESERT);
-    }
-
-    #[test]
-    fn wide_area_surface() {
-        use crate::generation::noise::router::multi_noise_sampler::{
-            MultiNoiseSampler, MultiNoiseSamplerBuilderOptions,
-        };
-        use crate::generation::{biome_coords, positions::chunk_pos};
-        #[derive(Deserialize)]
-        struct BiomeData {
-            x: i32,
-            z: i32,
-            data: Vec<(i32, i32, i32, u8)>,
-        }
-
-        let expected_data: Vec<BiomeData> =
-            read_data_from_file!("../../assets/biome_no_blend_no_beard_0.json");
-
-        let seed = 0;
-        use crate::generation::generator::{GeneratorInit, VanillaGenerator};
-        use pumpkin_util::world_seed::Seed;
-        let generator = VanillaGenerator::new(Seed(seed as u64), Dimension::OVERWORLD);
-
-        for data in expected_data {
-            let chunk_x = data.x;
-            let chunk_z = data.z;
-
-            let mut chunk = ProtoChunk::new(chunk_x, chunk_z, &generator);
-
-            // Create MultiNoiseSampler for populate_biomes
-
-            let start_x = chunk_pos::start_block_x(chunk_x);
-            let start_z = chunk_pos::start_block_z(chunk_z);
-
-            let horizontal_biome_end = biome_coords::from_block(16);
-            let multi_noise_config = MultiNoiseSamplerBuilderOptions::new(
-                biome_coords::from_block(start_x),
-                biome_coords::from_block(start_z),
-                horizontal_biome_end as usize,
-            );
-            let mut multi_noise_sampler = MultiNoiseSampler::generate(
-                &generator.base_router.multi_noise,
-                &multi_noise_config,
-            );
-
-            chunk.populate_biomes(&generator, &mut multi_noise_sampler);
-
-            for (biome_x, biome_y, biome_z, biome_id) in data.data {
-                let calculated_biome = chunk.get_biome(biome_x, biome_y, biome_z);
-
-                assert_eq!(
-                    biome_id,
-                    calculated_biome.id,
-                    "Expected {:?} was {:?} at {},{},{} ({},{})",
-                    Biome::from_id(biome_id),
-                    calculated_biome,
-                    biome_x,
-                    biome_y,
-                    biome_z,
-                    data.x,
-                    data.z
-                );
-            }
-        }
-    }
+    // Biome supplier tests that need `VanillaGenerator` (`biome_desert`,
+    // `wide_area_surface`) live in `pumpkin-worldgen/tests/biome.rs`.
 
     #[test]
     fn hash_seed_test() {
